use 'numeric' type for all 'wei' columns. last_incomplete_block -> first_incomplete_block. delete_last_incomplete_block -> delete_incomplete_blocks. added new views: crc_balances and safe_timeline(safe_address text)

+++ CirclesLand.BlockchainIndexer/Persistence/DetailWriters/CrcTransferWriter.cs
@@ -19,7 +19,7 @@ namespace CirclesLand.BlockchainIndexer.Persistence.DetailWriters
                     , ""to""
                     , value
                 ) values (
-                    @transaction_id, @from, @to, @value
+                    @transaction_id, @from, @to, @value::numeric
                 )
                 returning Id;
             ";
